+++ pagerduty/rest_api_v2_client.py
@@ -1,30 +1,28 @@
-
-# Copyright (c) PagerDuty.
-# See LICENSE for details.
-
-# Standard libraries
-import logging
-import sys
-import time
+# Core
 from copy import deepcopy
-from datetime import datetime
-from random import random
 from typing import Iterator, Union
 from warnings import warn
 
-# Upstream components on which this client is based:
-from requests import Response, Session
-from requests import __version__ as REQUESTS_VERSION
+# PyPI
+from requests import Response
 
-# HTTP client exceptions:
-from urllib3.exceptions import HTTPError, PoolError
-from requests.exceptions import RequestException
-
-__version__ = '1.0.0'
+# Local
+from . api_client import ApiClient, normalize_url
+from . common import (
+    auto_json,
+    requires_success,
+    singular_name,
+    successful_response,
+    truncate_text,
+    try_decoding,
+    ServerHttpError,
+    UrlError
+)
 
 #######################
 ### CLIENT DEFAULTS ###
 #######################
+
 ITERATION_LIMIT = 1e4
 """
 The maximum position of a result in classic pagination.
@@ -37,21 +35,7 @@ See: `Pagination
 <https://developer.pagerduty.com/docs/ZG9jOjExMDI5NTU4-pagination>`_.
 """
 
-TIMEOUT = 60
-"""
-The default timeout in seconds for any given HTTP request.
-
-Modifying this value will not affect any preexisting API session instances.
-Rather, it will only affect new instances. It is recommended to use
-:attr:`ApiClient.timeout` to configure the timeout for a given session.
-"""
-
-TEXT_LEN_LIMIT = 100
-"""
-The longest permissible length of API content to include in error messages.
-"""
-
-# List of canonical API paths
+# List of canonical REST API paths
 #
 # Supporting a new API for entity wrapping will require adding its patterns to
 # this list. If it doesn't follow standard naming conventions, it will also
@@ -454,9 +438,9 @@ properties in request bodies.
 """
 
 
-####################
-### URL HANDLING ###
-####################
+################################
+### REST API V2 URL HANDLING ###
+################################
 
 def canonical_path(base_url: str, url: str) -> str:
     """
@@ -544,32 +528,9 @@ def is_path_param(path_node: str) -> bool:
     """
     return path_node.startswith('{') and path_node.endswith('}')
 
-def normalize_url(base_url: str, url: str) -> str:
-    """
-    Normalize a URL or path to be a complete API URL before query parameters.
-
-    The ``url`` argument may be a path relative to the base URL or a full URL.
-
-    :param url:
-        The URL or path to normalize to a full URL.
-    :param base_url:
-        The base API URL, excluding any trailing slash, i.e.
-        "https://api.pagerduty.com"
-    :returns:
-        The full API URL.
-    """
-    if url.startswith(base_url):
-        return url
-    elif not (url.startswith('http://') or url.startswith('https://')):
-        return base_url.rstrip('/') + "/" + url.lstrip('/')
-    else:
-        raise UrlError(
-            f"URL {url} does not start with the API base URL {base_url}"
-        )
-
-#######################
-### ENTITY WRAPPING ###
-#######################
+###############################
+### ENTITY WRAPPING HELPERS ###
+###############################
 
 def entity_wrappers(method: str, path: str) -> tuple:
     """
@@ -690,29 +651,6 @@ def unwrap(response: Response, wrapper) -> Union[dict, list]:
 ### FUNCTION DECORATORS ###
 ###########################
 
-def auto_json(method):
-    """
-    Makes methods return the full response body object after decoding from JSON.
-
-    Intended for use on functions that take a URL positional argument followed
-    by keyword arguments and return a `requests.Response`_ object.
-    """
-    doc = method.__doc__
-    def call(self, url, **kw):
-        return try_decoding(successful_response(method(self, url, **kw)))
-    call.__doc__ = doc
-    return call
-
-def requires_success(method):
-    """
-    Decorator that validates HTTP responses.
-    """
-    doc = method.__doc__
-    def call(self, url, **kw):
-        return successful_response(method(self, url, **kw))
-    call.__doc__ = doc
-    return call
-
 def resource_url(method):
     """
     API call decorator that allows passing a resource dict as the path/URL
@@ -792,796 +730,9 @@ def wrapped_entities(method):
     call.__doc__ = doc
     return call
 
-
-########################
-### HELPER FUNCTIONS ###
-########################
-
-def deprecated_kwarg(deprecated_name: str, details=None, method=None):
-    """
-    Raises a warning if a deprecated keyword argument is used.
-
-    :param deprecated_name: The name of the deprecated function
-    :param details: An optional message to append to the deprecation message
-    :param method: An optional method name
-    """
-    details_msg = ''
-    method_msg = ''
-    if method is not None:
-        method_msg = f" of {method}"
-    if details is not None:
-        details_msg = f" {details}"
-    warn(
-        f"Keyword argument \"{deprecated_name}\"{method_msg} is deprecated."+details_msg
-    )
-
-def http_error_message(r: Response, context=None) -> str:
-    """
-    Formats a message describing a HTTP error.
-
-    :param r:
-        The response object.
-    :param context:
-        A description of when the error was received, or None to not include it
-    :returns:
-        The message to include in the HTTP error
-    """
-    received_http_response = bool(r.status_code)
-    endpoint = "%s %s"%(r.request.method.upper(), r.request.url)
-    context_msg = ""
-    if type(context) is str:
-        context_msg=f" in {context}"
-    if received_http_response and not r.ok:
-        err_type = 'unknown'
-        if r.status_code / 100 == 4:
-            err_type = 'client'
-        elif r.status_code / 100 == 5:
-            err_type = 'server'
-        tr_bod = truncate_text(r.text)
-        return f"{endpoint}: API responded with {err_type} error (status " \
-            f"{r.status_code}){context_msg}: {tr_bod}"
-    elif not received_http_response:
-        return f"{endpoint}: Network or other unknown error{context_msg}"
-    else:
-        return f"{endpoint}: Success (status {r.status_code}) but an " \
-            f"expectation still failed{context_msg}"
-
-def last_4(secret: str) -> str:
-    """
-    Truncate a sensitive value to its last 4 characters
-
-    :param secret: text to truncate
-    :returns:
-        The truncated text
-    """
-    return '*'+str(secret)[-4:]
-
-def plural_name(obj_type: str) -> str:
-    """
-    Pluralizes a name, i.e. the API name from the ``type`` property
-
-    :param obj_type:
-        The object type, i.e. ``user`` or ``user_reference``
-    :returns:
-        The name of the resource, i.e. the last part of the URL for the
-        resource's index URL
-    """
-    if obj_type.endswith('_reference'):
-        # Strip down to basic type if it's a reference
-        obj_type = obj_type[:obj_type.index('_reference')]
-    if obj_type.endswith('y'):
-        # Because English
-        return obj_type[:-1]+'ies'
-    else:
-        return obj_type+'s'
-
-def singular_name(r_name: str) -> str:
-    """
-    Singularizes a name, i.e. for the entity wrapper in a POST request
-
-    :para r_name:
-        The "resource" name, i.e. "escalation_policies", a plural noun that
-        forms the part of the canonical path identifying what kind of resource
-        lives in the collection there, for an API that follows classic wrapped
-        entity naming patterns.
-    :returns:
-        The singularized name
-    """
-    if r_name.endswith('ies'):
-        # Because English
-        return r_name[:-3]+'y'
-    else:
-        return r_name.rstrip('s')
-
-def successful_response(r: Response, context=None) -> Response:
-    """Validates the response as successful.
-
-    Returns the response if it was successful; otherwise, raises an exception.
-
-    :param r:
-        Response object corresponding to the response received.
-    :param context:
-        A description of when the HTTP request is happening, for error reporting
-    :returns:
-        The response object, if it was successful
-    """
-    if r.ok and bool(r.status_code):
-        return r
-    elif r.status_code / 100 == 5:
-        raise ServerHttpError(http_error_message(r, context=context), r)
-    elif bool(r.status_code):
-        raise HttpError(http_error_message(r, context=context), r)
-    else:
-        raise Error(http_error_message(r, context=context))
-
-def truncate_text(text: str) -> str:
-    """Truncates a string longer than :attr:`TEXT_LEN_LIMIT`
-
-    :param text: The string to truncate if longer than the limit.
-    """
-    if len(text) > TEXT_LEN_LIMIT:
-        return text[:TEXT_LEN_LIMIT-1]+'...'
-    else:
-        return text
-
-def try_decoding(r: Response) -> Union[dict, list, str]:
-    """
-    JSON-decode a response body
-
-    Returns the decoded body if successful; raises :class:`ServerHttpError`
-    otherwise.
-
-    :param r:
-        The response object
-    """
-    try:
-        return r.json()
-    except ValueError as e:
-        raise ServerHttpError(
-            "API responded with invalid JSON: " + truncate_text(r.text),
-            r,
-        )
-
-###############
-### CLASSES ###
-###############
-
-class ApiClient(Session):
-    """
-    Base class for making HTTP requests to PagerDuty APIs
-
-    This is an opinionated wrapper of `requests.Session`_, with a few additional
-    features:
-
-    - The client will reattempt the request with auto-increasing cooldown/retry
-      intervals, with attempt limits configurable through the :attr:`retry`
-      attribute.
-    - When making requests, headers specified ad-hoc in calls to HTTP verb
-      functions will not replace, but will be merged into, default headers.
-    - The request URL, if it doesn't already start with the REST API base URL,
-      will be prepended with the default REST API base URL.
-    - It will only perform requests with methods as given in the
-      :attr:`permitted_methods` list, and will raise :class:`Error` for
-      any other HTTP methods.
-
-    :param api_key:
-        REST API access token to use for HTTP requests
-    :param debug:
-        Sets :attr:`print_debug`. Set to True to enable verbose command line
-        output.
-    :type token: str
-    :type debug: bool
-    """
-
-    log = None
-    """
-    A ``logging.Logger`` object for logging messages. By default it is
-    configured without any handlers and so no messages will be emitted. See:
-    `Logger Objects
-    <https://docs.python.org/3/library/logging.html#logger-objects>`_.
-    """
-
-    max_http_attempts = 10
-    """
-    The number of times that the client will retry after error statuses, for any
-    that are defined greater than zero in :attr:`retry`.
-    """
-
-    max_network_attempts = 3
-    """
-    The number of times that connecting to the API will be attempted before
-    treating the failure as non-transient; a :class:`Error` exception
-    will be raised if this happens.
-    """
-
-    parent = None
-    """The ``super`` object (`requests.Session`_)"""
-
-    permitted_methods = ()
-    """
-    A tuple of the methods permitted by the API which the client implements.
-
-    For instance:
-
-    * The REST API accepts GET, POST, PUT and DELETE.
-    * The Events API and Change Events APIs only accept POST.
-    """
-
-    retry = {}
-    """
-    A dict defining the retry behavior for each HTTP response status code.
-
-    Each key in this dictionary is an int representing a HTTP response code. The
-    behavior is specified by the int value at each key as follows:
-
-    * ``-1`` to retry without limit.
-    * ``0`` has no effect; the default behavior will take effect.
-    * ``n``, where ``n > 0``, to retry ``n`` times (or up
-      to :attr:`max_http_attempts` total for all statuses, whichever is
-      encountered first), and then return the final response.
-
-    The default behavior is to retry without limit on status 429, raise an
-    exception on a 401, and return the `requests.Response`_ object in any other case
-    (assuming a HTTP response was received from the server).
-    """
-
-    sleep_timer = 1.5
-    """
-    Default initial cooldown time factor for rate limiting and network errors.
-
-    Each time that the request makes a followup request, there will be a delay
-    in seconds equal to this number times :attr:`sleep_timer_base` to the power
-    of how many attempts have already been made so far, unless
-    :attr:`stagger_cooldown` is nonzero.
-    """
-
-    sleep_timer_base = 2
-    """
-    After each retry, the time to sleep before reattempting the API connection
-    and request will increase by a factor of this amount.
-    """
-
-    timeout = TIMEOUT
-    """
-    This is the value sent to `Requests`_ as the ``timeout`` parameter that
-    determines the TCP read timeout.
-    """
-
-    url = ""
-
-    def __init__(self, api_key: str, debug=False):
-        self.parent = super(ApiClient, self)
-        self.parent.__init__()
-        self.api_key = api_key
-        self.log = logging.getLogger(__name__)
-        self.print_debug = debug
-        self.retry = {}
-
-    def after_set_api_key(self):
-        """
-        Setter hook for setting or updating the API key.
-
-        Child classes should implement this to perform additional steps.
-        """
-        pass
-
-    @property
-    def api_key(self) -> str:
-        """
-        Property representing the credential used for accessing the given API.
-        """
-        return self._api_key
-
-    @api_key.setter
-    def api_key(self, api_key):
-        if not (isinstance(api_key, str) and api_key):
-            raise ValueError("API credential must be a non-empty string.")
-        self._api_key = api_key
-        self.headers.update(self.auth_header)
-        self.after_set_api_key()
-
-    @property
-    def auth_header(self) -> dict:
-        """
-        Generates the header with the API credential used for authentication.
-        """
-        raise NotImplementedError
-
-    def cooldown_factor(self) -> float:
-        return self.sleep_timer_base*(1+self.stagger_cooldown*random())
-
-    def normalize_params(self, params) -> dict:
-        """
-        Modify the user-supplied parameters to ease implementation
-
-        Current behavior:
-
-        * If a parameter's value is of type list, and the parameter name does
-          not already end in "[]", then the square brackets are appended to keep
-          in line with the requirement that all set filters' parameter names end
-          in "[]".
-
-        :returns:
-            The query parameters after modification
-        """
-        updated_params = {}
-        for param, value in params.items():
-            if type(value) is list and not param.endswith('[]'):
-                updated_params[param+'[]'] = value
-            else:
-                updated_params[param] = value
-        return updated_params
-
-    def normalize_url(self, url) -> str:
-        """Compose the URL whether it is a path or an already-complete URL"""
-        return normalize_url(self.url, url)
-
-    def postprocess(self, response):
-        """
-        Perform supplemental actions immediately after receiving a response.
-
-        This method is called once per request not including retries, and can be
-        extended in child classes.
-        """
-        pass
-
-    def prepare_headers(self, method, user_headers={}) -> dict:
-        """
-        Append special additional per-request headers.
-
-        :param method:
-            The HTTP method, in upper case.
-        :param user_headers:
-            Headers that can be specified to override default values.
-        :returns:
-            The final list of headers to use in the request
-        """
-        headers = deepcopy(self.headers)
-        if user_headers:
-            headers.update(user_headers)
-        return headers
-
-    @property
-    def print_debug(self) -> bool:
-        """
-        Printing debug flag
-
-        If set to True, the logging level of :attr:`log` is set to
-        ``logging.DEBUG`` and all log messages are emitted to ``sys.stderr``.
-        If set to False, the logging level of :attr:`log` is set to
-        ``logging.NOTSET`` and the debugging log handler that prints messages to
-        ``sys.stderr`` is removed. This value thus can be toggled to enable and
-        disable verbose command line output.
-
-        It is ``False`` by default and it is recommended to keep it that way in
-        production settings.
-        """
-        return self._debug
-
-    @print_debug.setter
-    def print_debug(self, debug: bool):
-        self._debug = debug
-        if debug and not hasattr(self, '_debugHandler'):
-            self.log.setLevel(logging.DEBUG)
-            self._debugHandler = logging.StreamHandler()
-            self.log.addHandler(self._debugHandler)
-        elif not debug and hasattr(self, '_debugHandler'):
-            self.log.setLevel(logging.NOTSET)
-            self.log.removeHandler(self._debugHandler)
-            delattr(self, '_debugHandler')
-        # else: no-op; only happens if debug is set to the same value twice
-
-    def request(self, method, url, **kwargs) -> Response:
-        """
-        Make a generic PagerDuty API request.
-
-        :param method:
-            The request method to use. Case-insensitive. May be one of get, put,
-            post or delete.
-        :param url:
-            The path/URL to request. If it does not start with the base URL, the
-            base URL will be prepended.
-        :param **kwargs:
-            Custom keyword arguments to pass to ``requests.Session.request``.
-        :type method: str
-        :type url: str
-        :returns:
-            The `requests.Response`_ object corresponding to the HTTP response
-        """
-        sleep_timer = self.sleep_timer
-        network_attempts = 0
-        http_attempts = {}
-        method = method.strip().upper()
-        if method not in self.permitted_methods:
-            m_str = ', '.join(self.permitted_methods)
-            raise Error(f"Method {method} not supported by this API. " \
-                f"Permitted methods: {m_str}")
-        req_kw = deepcopy(kwargs)
-        full_url = self.normalize_url(url)
-        endpoint = "%s %s"%(method.upper(), full_url)
-
-        # Add in any headers specified in keyword arguments:
-        headers = kwargs.get('headers', {})
-        req_kw.update({
-            'headers': self.prepare_headers(method, user_headers=headers),
-            'stream': False,
-            'timeout': self.timeout
-        })
-
-        # Special changes to user-supplied parameters, for convenience
-        if 'params' in kwargs and kwargs['params']:
-            req_kw['params'] = self.normalize_params(kwargs['params'])
-
-        # Make the request (and repeat w/cooldown if the rate limit is reached):
-        while True:
-            try:
-                response = self.parent.request(method, full_url, **req_kw)
-                self.postprocess(response)
-            except (HTTPError, PoolError, RequestException) as e:
-                network_attempts += 1
-                if network_attempts > self.max_network_attempts:
-                    error_msg = f"{endpoint}: Non-transient network " \
-                        'error; exceeded maximum number of attempts ' \
-                        f"({self.max_network_attempts}) to connect to the API."
-                    raise Error(error_msg) from e
-                sleep_timer *= self.cooldown_factor()
-                self.log.warning(
-                    "%s: HTTP or network error: %s. retrying in %g seconds.",
-                    endpoint, e.__class__.__name__, sleep_timer)
-                time.sleep(sleep_timer)
-                continue
-
-            status = response.status_code
-            retry_logic = self.retry.get(status, 0)
-            if not response.ok and retry_logic != 0:
-                # Take special action as defined by the retry logic
-                if retry_logic != -1:
-                    # Retry a specific number of times (-1 implies infinite)
-                    if http_attempts.get(status, 0)>=retry_logic or \
-                            sum(http_attempts.values())>self.max_http_attempts:
-                        lower_limit = retry_logic
-                        if lower_limit > self.max_http_attempts:
-                            lower_limit = self.max_http_attempts
-                        self.log.error(
-                            f"%s: Non-transient HTTP error: exceeded " \
-                            'maximum number of attempts (%d) to make a ' \
-                            'successful request. Currently encountering ' \
-                            'status %d.', endpoint, lower_limit, status)
-                        return response
-                    http_attempts[status] = 1 + http_attempts.get(status, 0)
-                sleep_timer *= self.cooldown_factor()
-                self.log.warning("%s: HTTP error (%d); retrying in %g seconds.",
-                    endpoint, status, sleep_timer)
-                time.sleep(sleep_timer)
-                continue
-            elif status == 429:
-                sleep_timer *= self.cooldown_factor()
-                self.log.debug("%s: Hit API rate limit (status 429); " \
-                    "retrying in %g seconds", endpoint, sleep_timer)
-                time.sleep(sleep_timer)
-                continue
-            elif status == 401:
-                # Stop. Authentication failed. We shouldn't try doing any more,
-                # because we'll run into the same problem later anyway.
-                raise HttpError(
-                    "Received 401 Unauthorized response from the API. The key "
-                    "(...%s) may be invalid or deactivated."%self.trunc_key,
-                    response)
-            else:
-                # All went according to plan.
-                return response
-
-    @property
-    def stagger_cooldown(self) -> float:
-        """
-        Randomizing factor for wait times between retries during rate limiting.
-
-        If set to number greater than 0, the sleep time for rate limiting will
-        (for each successive sleep) be adjusted by a factor of one plus a
-        uniformly-distributed random number between 0 and 1 times this number,
-        on top of the base sleep timer :attr:`sleep_timer_base`.
-
-        For example:
-
-        * If this is 1, and :attr:`sleep_timer_base` is 2 (default), then after
-          each status 429 response, the sleep time will change overall by a
-          random factor between 2 and 4, whereas if it is zero, it will change
-          by a factor of 2.
-        * If :attr:`sleep_timer_base` is 1, then the cooldown time will be
-          adjusted by a random factor between one and one plus this number.
-
-        If the number is set to zero, then this behavior is effectively
-        disabled, and the cooldown factor (by which the sleep time is adjusted)
-        will just be :attr:`sleep_timer_base`
-
-        Setting this to a nonzero number helps avoid the "thundering herd"
-        effect that can potentially be caused by many API clients making
-        simultaneous concurrent API requests and consequently waiting for the
-        same amount of time before retrying.  It is currently zero by default
-        for consistent behavior with previous versions.
-        """
-        if hasattr(self, '_stagger_cooldown'):
-            return self._stagger_cooldown
-        else:
-            return 0
-
-    @stagger_cooldown.setter
-    def stagger_cooldown(self, val):
-        if type(val) not in [float, int] or val<0:
-            raise ValueError("Cooldown randomization factor stagger_cooldown "
-                "must be a positive real number")
-        self._stagger_cooldown = val
-
-    @property
-    def trunc_key(self) -> str:
-        """Truncated key for secure display/identification purposes."""
-        return last_4(self.api_key)
-
-    @property
-    def user_agent(self) -> str:
-        return 'pagerduty/%s python-requests/%s Python/%d.%d'%(
-            __version__,
-            REQUESTS_VERSION,
-            sys.version_info.major,
-            sys.version_info.minor
-        )
-
-class EventsApiV2Client(ApiClient):
-
-    """
-    Session class for submitting events to the PagerDuty v2 Events API.
-
-    Implements methods for submitting events to PagerDuty through the Events API,
-    including change events, and inherits from :class:`pagerduty.ApiClient`.  For more
-    details on usage of this API, refer to the `Events API v2 documentation
-    <https://developer.pagerduty.com/docs/events-api-v2/overview/>`_
-    """
-
-    permitted_methods = ('POST',)
-
-    url = "https://events.pagerduty.com"
-
-    def __init__(self, api_key: str, debug=False):
-        super(EventsApiV2Client, self).__init__(api_key, debug)
-        # See: https://developer.pagerduty.com/docs/3d063fd4814a6-events-api-v2-overview#response-codes--retry-logic
-        self.retry[500] = 2 # internal server error
-        self.retry[502] = 4 # bad gateway
-        self.retry[503] = 6 # service unavailable
-        self.retry[504] = 6 # gateway timeout
-
-    @property
-    def auth_header(self) -> dict:
-        return {}
-
-    def acknowledge(self, dedup_key) -> str:
-        """
-        Acknowledge an alert via Events API.
-
-        :param dedup_key:
-            The deduplication key of the alert to set to the acknowledged state.
-        :returns:
-            The deduplication key
-        """
-        return self.send_event('acknowledge', dedup_key=dedup_key)
-
-    @property
-    def event_timestamp(self) -> str:
-        return datetime.utcnow().isoformat()+'Z'
-
-    def post(self, *args, **kw) -> Response:
-        """
-        Override of ``requests.Session.post``
-
-        Adds the ``routing_key`` parameter to the body before sending.
-        """
-        if 'json' in kw and hasattr(kw['json'], 'update'):
-            kw['json'].update({'routing_key': self.api_key})
-        return super(EventsApiV2Client, self).post(*args, **kw)
-
-    def prepare_headers(self, method, user_headers={}) -> dict:
-        """
-        Add user agent and content type headers for Events API requests.
-
-        :param user_headers: User-supplied headers that will override defaults
-        :returns:
-            The final list of headers to use in the request
-        """
-        headers = {}
-        headers.update(self.headers)
-        headers.update({
-            'Content-Type': 'application/json',
-            'User-Agent': self.user_agent,
-        })
-        headers.update(user_headers)
-        return headers
-
-    def resolve(self, dedup_key) -> str:
-        """
-        Resolve an alert via Events API.
-
-        :param dedup_key:
-            The deduplication key of the alert to resolve.
-        """
-        return self.send_event('resolve', dedup_key=dedup_key)
-
-    def send_change_event(self, payload={}, links=[], routing_key=None):
-        """
-        Send a change event to the v2 Change Events API.
-
-        See: https://developer.pagerduty.com/docs/events-api-v2/send-change-events/
-
-        :param payload:
-            A dictionary object with keys ``summary``, ``source``, ``timestamp`` and
-            ``custom_details`` as described in the above documentation.
-        :param links:
-            A list of dictionary objects each with keys ``href`` and ``text``
-            representing the target and display text of each link
-        :param routing_key:
-            (Deprecated) the routing key. The parameter is set automatically to the
-            :attr:`ApiClient.api_key` property in the final payload and this argument is ignored.
-        :returns:
-            The response ID
-        """
-        if routing_key is not None:
-            deprecated_kwarg(
-                'routing_key',
-                method='EventsApiV2Client.send_change_event'
-            )
-        event = {'payload': deepcopy(payload)}
-        if links:
-            event['links'] = deepcopy(links)
-        response = self.post('/v2/change/enqueue', json=event)
-        response_body = try_decoding(successful_response(
-            response,
-            context="submitting change event",
-        ))
-
-    def send_event(self, action, dedup_key=None, **properties) -> str:
-        """
-        Send an event to the v2 Events API.
-
-        See: https://v2.developer.pagerduty.com/docs/send-an-event-events-api-v2
-
-        :param action:
-            The action to perform through the Events API: trigger, acknowledge
-            or resolve.
-        :param dedup_key:
-            The deduplication key; used for determining event uniqueness and
-            associating actions with existing incidents.
-        :param **properties:
-            Additional properties to set, i.e. if ``action`` is ``trigger``
-            this would include ``payload``.
-        :type action: str
-        :type dedup_key: str
-        :returns:
-            The deduplication key of the incident
-        """
-
-        actions = ('trigger', 'acknowledge', 'resolve')
-        if action not in actions:
-            raise ValueError("Event action must be one of: "+', '.join(actions))
-
-        event = {'event_action':action}
-
-        event.update(properties)
-        if isinstance(dedup_key, str):
-            event['dedup_key'] = dedup_key
-        elif not action == 'trigger':
-            raise ValueError("The dedup_key property is required for"
-                "event_action=%s events, and it must be a string."%action)
-        response = successful_response(
-            self.post('/v2/enqueue', json=event),
-            context='submitting an event to the events API',
-        )
-        response_body = try_decoding(response)
-        if type(response_body) is not dict or 'dedup_key' not in response_body:
-            err_msg = 'Malformed response body from the events API; it is ' \
-                'not a dict that has a key named "dedup_key" after ' \
-                'decoding. Body = '+truncate_text(response.text)
-            raise ServerHttpError(err_msg, response)
-        return response_body['dedup_key']
-
-    def submit(self, summary, source=None, custom_details=None, links=None,
-            timestamp=None) -> str:
-        """
-        Submit a change event.
-
-        See: https://developer.pagerduty.com/docs/send-change-event
-
-        :param summary:
-            Summary / brief description of the change, for ``payload.summary``.
-        :param source:
-            A human-readable name identifying the source of the change, for the
-            ``payload.source`` event property.
-        :param custom_details:
-            A dictionary object to use as the ``payload.custom_details`` property.
-        :param links:
-            A list of dict objects to use as the ``links`` property of the event.
-        :param timestamp:
-            Specifies an event timestamp. Must be an ISO8601-format date/time.
-        :type summary: str
-        :type source: str
-        :type custom_details: dict
-        :type links: list
-        :type timestamp: str
-        :returns:
-            The response ID
-        """
-        local_var = locals()['custom_details']
-        if not (local_var is None or isinstance(local_var, dict)):
-            raise ValueError("custom_details must be a dict")
-        if timestamp is None:
-            timestamp = self.event_timestamp
-        event = {
-                'payload': {
-                    'summary': summary,
-                    'timestamp': timestamp,
-                    }
-                }
-        if isinstance(source, str):
-            event['payload']['source'] = source
-        if isinstance(custom_details, dict):
-            event['payload']['custom_details'] = custom_details
-        if links:
-            event['links'] = links
-        return self.send_change_event(**event)
-
-    def trigger(self, summary, source, dedup_key=None, severity='critical',
-            payload=None, custom_details=None, images=None, links=None) -> str:
-        """
-        Send an alert-triggering event
-
-        :param summary:
-            Summary / brief description of what is wrong.
-        :param source:
-            A human-readable name identifying the system that is affected.
-        :param dedup_key:
-            The deduplication key; used for determining event uniqueness and
-            associating actions with existing incidents.
-        :param severity:
-            Alert severity. Sets the ``payload.severity`` property.
-        :param payload:
-            Set the payload directly. Can be used in conjunction with other
-            parameters that also set payload properties; these properties will
-            be merged into the default payload, and any properties in this
-            parameter will take precedence except with regard to
-            ``custom_details``.
-        :param custom_details:
-            The ``payload.custom_details`` property of the payload. Will
-            override the property set in the ``payload`` parameter if given.
-        :param images:
-            Set the ``images`` property of the event.
-        :param links:
-            Set the ``links`` property of the event.
-        :type action: str
-        :type custom_details: dict
-        :type dedup_key: str
-        :type images: list
-        :type links: list
-        :type payload: dict
-        :type severity: str
-        :type source: str
-        :type summary: str
-        :returns:
-            The deduplication key of the incident, if any.
-        """
-        for local in ('payload', 'custom_details'):
-            local_var = locals()[local]
-            if not (local_var is None or type(local_var) is dict):
-                raise ValueError(local+" must be a dict")
-        event = {'payload': {'summary':summary, 'source':source,
-            'severity':severity}}
-        if type(payload) is dict:
-            event['payload'].update(payload)
-        if type(custom_details) is dict:
-            details = event.setdefault('payload', {}).get('custom_details', {})
-            details.update(custom_details)
-            event['payload']['custom_details'] = details
-        if images:
-            event['images'] = images
-        if links:
-            event['links'] = links
-        return self.send_event('trigger', dedup_key=dedup_key, **event)
+####################
+### CLIENT CLASS ###
+####################
 
 class RestApiV2Client(ApiClient):
     """
@@ -2211,78 +1362,4 @@ class RestApiV2Client(ApiClient):
     @property
     def trunc_token(self) -> str:
         """Truncated token for secure display/identification purposes."""
-        return last_4(self.api_key)
-
-class UrlError(Exception):
-    """
-    Exception class for unsupported URLs or malformed input.
-    """
-    pass
-
-class Error(Exception):
-    """
-    General API errors base class.
-
-    Note, the name of this class does not imply it solely includes errors
-    experienced by the client or HTTP status 4xx responses, but descendants can
-    include issues with the API backend.
-    """
-
-    response = None
-    """
-    The HTTP response object, if a response was successfully received.
-
-    In the case of network errors, this property will be None.
-    """
-
-    def __init__(self, message, response=None):
-        self.msg = message
-        self.response = response
-        super(Error, self).__init__(message)
-
-class HttpError(Error):
-    """
-    Error class representing errors strictly associated with HTTP responses.
-
-    This class was created to make it easier to more cleanly handle errors by
-    way of a class that is guaranteed to have its ``response`` be a valid
-    `requests.Response`_ object.
-
-    Whereas, the more generic :class:`Error` could also be used
-    to denote such things as non-transient network errors wherein no response
-    was recevied from the API.
-
-    For instance, instead of this:
-
-    ::
-
-        try:
-            user = session.rget('/users/PABC123')
-        except pagerduty.Error as e:
-            if e.response is not None:
-                print("HTTP error: "+str(e.response.status_code))
-            else:
-                raise e
-
-    one could write this:
-
-    ::
-
-        try:
-            user = session.rget('/users/PABC123')
-        except pagerduty.HttpError as e:
-            print("HTTP error: "+str(e.response.status_code))
-    """
-
-    def __init__(self, message, response: Response):
-        super(HttpError, self).__init__(message, response=response)
-
-class ServerHttpError(HttpError):
-    """
-    Error class representing failed expectations made of the server.
-
-    This is raised in cases where the response schema differs from the expected schema
-    because of an API bug, or because it's an early access endpoint and changes before
-    GA, or in cases of HTTP status 5xx where a successful response is required.
-    """
-    pass
+        return self.trunc_key
