+++ pyproject.toml
@@ -1,0 +1,7 @@
+[project]
+name = "pagerduty"
+version = "2.0.0"
+
+[build-system]
+requires = ["setuptools", "requests", "urllib3"]
+build-backend = "setuptools.build_meta"
